services/data-ingest/requirements.txt
dependency-injector==4.36.0
kafka-python==2.0.2
python-dotenv==0.19.0
PyYAML==5.4.1
mariadb==1.0.7
pandas==1.3.3
